+++ client/src/app.js
@@ -12,7 +12,8 @@ import Dashboard from "./containers/dashboard";
 import Home from "./containers/home";
 import Account from "./containers/account";
 import Profile from "./containers/profile";
-import Survey from "./containers/survey";
+import Jobs from "./containers/jobs";
+import NewJob from "./containers/newJob";
 import Blog from "./containers/blog";
 
 import styles from "./styles/global.scss";
@@ -23,7 +24,7 @@ class App extends Component {
     this.props.fetchUser();
   }
   componentDidUpdate() {
-    // ensuring any updates asynchrnously update our credits
+    // ensuring any updates asynchronously update our credits
     this.props.fetchUser();
   }
   render() {
@@ -33,10 +34,11 @@ class App extends Component {
         <main className="container">
           <Route exact path="/" component={Home} />
           <Route exact path="/dashboard" component={Dashboard} />
-          <Route path="/dashboard/account" component={Account} />
-          <Route path="/dashboard/profile" component={Profile} />
-          <Route path="/dashboard/survey" component={Survey} />
-          <Route path="/dashboard/blog" component={Blog} />
+            <Route path="/dashboard/account" component={Account} />
+            <Route path="/dashboard/profile" component={Profile} />
+            <Route exact path="/dashboard/jobs" component={Jobs} />
+              <Route path="/dashboard/jobs/new" component={NewJob} />
+            <Route path="/dashboard/blog" component={Blog} />
         </main>
         <Footer />
       </div>