+++ client/src/reducers/index.js
@@ -1,7 +1,12 @@
 import { combineReducers } from "redux";
 
+import {reducer as reduxForm } from "redux-form"
+
 import authReducer from "./authReducer";
 
 export default combineReducers({
-	auth: authReducer
+	auth: authReducer,
+
+	// this is necessary to globally utilize the redux-form state
+	form: reduxForm
 });